Replace cbor_slides banner to deprecate v1
This is a png because gif with transparent backgound wasn't good for
text readability.  Old .gif filename is used to avoid changing URL in 
README and tagging new release to make it show up in pkg.go.dev.

png works well for single image with text to be readable on light and dark 
backgounds.